feat(igxDrag/igxDrop): Rename main files and folder to reflect Angular style guide. #5319
