+++ projects/igniteui-angular/src/lib/directives/drag-drop/drag-drop.directive.ts
