+++ projects/igniteui-angular/src/lib/grids/grid-common.module.ts
@@ -39,7 +39,7 @@ import {
 import { IgxGridTransaction } from './grid-base.component';
 import { IgxRowComponent } from './row.component';
 import { IgxChipsModule } from '../chips/chips.module';
-import { IgxDragDropModule } from '../directives/dragdrop/dragdrop.directive';
+import { IgxDragDropModule } from '../directives/drag-drop/drag-drop.directive';
 import { IgxButtonGroupModule } from '../buttonGroup/buttonGroup.component';
 import { IgxColumnPinningModule } from './column-pinning.component';
 import { IgxBaseTransactionService } from '../services/transaction/base-transaction';
